# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Comfort taxi API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/order": {
            "get": {
                "tags": [
                    "order"
                ],
                "summary": "Get list of order",
                "operationId": "b459854003e608556bb3cb6fea8411b0",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized user"
                    },
                    "400": {
                        "description": "order isn't found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/order/{id}": {
            "get": {
                "tags": [
                    "order"
                ],
                "summary": "Get order by id",
                "description": "Get order by id",
                "operationId": "getorder",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of order class",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "order": {
                                            "$ref": "#/components/schemas/OrderResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "order is not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "order is not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized user",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "order"
                ],
                "summary": "Create order",
                "description": "Create new order",
                "operationId": "createorder",
                "requestBody": {
                    "description": "Send data to create a new order",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "user": {
                                        "$ref": "#/components/schemas/OrderCreateRequest"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "task": {
                                            "$ref": "#/components/schemas/OrderResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized user",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error when creating order",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Error when creating order"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "order"
                ],
                "summary": "Delete order by id",
                "description": "Delete order by id",
                "operationId": "deleteorder",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of order",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "order": {
                                            "$ref": "#/components/schemas/OrderResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "order is not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "order class is not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized user",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "order"
                ],
                "summary": "Update order",
                "description": "Update order",
                "operationId": "updateorder",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of order",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "requestBody": {
                    "description": "Send data to update a order",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "order": {
                                        "$ref": "#/components/schemas/OrderUpdateRequest"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "order": {
                                            "$ref": "#/components/schemas/OrderResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "order is not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "order is not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized user",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error when updating order",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Error when updating order"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "OrderCreateRequest": {
                "required": [
                    "status",
                    "working_shirt_id",
                    "lead_id",
                    "price",
                    "rating",
                    "date_start"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "status": {
                        "type": "string",
                        "example": "Выполняется"
                    },
                    "working_shirt_id": {
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "lead_id": {
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "price": {
                        "type": "biginteger",
                        "example": "1200"
                    },
                    "rating": {
                        "type": "number",
                        "format": "double",
                        "example": "4.8"
                    },
                    "date_start": {
                        "type": "string",
                        "format": "date",
                        "readOnly": "true",
                        "example": "2019-02-25"
                    },
                    "date_end": {
                        "type": "string",
                        "format": "date",
                        "readOnly": "true",
                        "example": "2019-02-25"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "OrderCreateRequest"
                }
            },
            "OrderUpdateRequest": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "status": {
                        "type": "string",
                        "example": "Выполняется"
                    },
                    "working_shirt_id": {
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "lead_id": {
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "price": {
                        "type": "biginteger",
                        "example": "1200"
                    },
                    "rating": {
                        "type": "number",
                        "format": "double",
                        "example": "4.8"
                    },
                    "date_start": {
                        "type": "string",
                        "format": "date",
                        "readOnly": "true",
                        "example": "2019-02-25"
                    },
                    "date_end": {
                        "type": "string",
                        "format": "date",
                        "readOnly": "true",
                        "example": "2019-02-25"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "OrderUpdateRequest"
                }
            },
...
            "OrderResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "status": {
                        "type": "string",
                        "example": "Выполняется"
                    },
                    "working_shirt_id": {
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "lead_id": {
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "price": {
                        "type": "biginteger",
                        "example": "1200"
                    },
                    "rating": {
                        "type": "number",
                        "format": "double",
                        "example": "4.8"
                    },
                    "date_start": {
                        "type": "string",
                        "format": "date",
                        "readOnly": "true",
                        "example": "2019-02-25"
                    },
                    "date_end": {
                        "type": "string",
                        "format": "date",
                        "readOnly": "true",
                        "example": "2019-02-25"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "OrderResource"
                }
            },
...
        },
...
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    ]
}